I translated the words into English. I discovered that Laravel comes alone with "axios" and that it sends a bad parameter that the Google API does not accept, so I deleted from the "welcome" page that script that loads "axios" "manually and I disabled that default setting in Laravel.
I fixed a bug in which the map and the search bar did not appear on the main page. More precisely, I manually loaded the function when the page loads, in the JS file at the end.
I added the Log Out button
We have added the functionality of displaying the search bar only if you are logged in to the site with an account.